Move EndpointCommand handling to crate::commands::endpoint to clean up the projection module
