+++ workspaces/diff-engine/src/projections/endpoint.rs
@@ -347,212 +347,3 @@ impl AggregateEvent<EndpointProjection> for SpecEvent {
     }
   }
 }
-
-impl AggregateCommand<EndpointProjection> for EndpointCommand {
-  type Error = SpecCommandError;
-  type Event = EndpointEvent;
-  type Events = Vec<EndpointEvent>;
-
-  fn execute_on(self, projection: &EndpointProjection) -> Result<Self::Events, Self::Error> {
-    let validation = CommandValidationQueries::from((projection, &self));
-
-    let events = match self {
-      // Path components
-      // ---------------
-      EndpointCommand::AddPathComponent(command) => {
-        validation.require(
-          validation.path_component_id_exists(&command.parent_path_id),
-          "parent path component must exist to add path component",
-        )?;
-        validation.require(
-          !validation.path_component_id_exists(&command.path_id),
-          "path id must be assignable to add path component",
-        )?;
-
-        vec![EndpointEvent::from(
-          endpoint_events::PathComponentAdded::from(command),
-        )]
-      }
-
-      EndpointCommand::RenamePathComponent(command) => {
-        validation.require(
-          !validation.path_component_is_root(&command.path_id),
-          "path id can not be root to rename path component",
-        )?;
-
-        validation.require(
-          validation.path_component_id_exists(&command.path_id),
-          "path component must exist to rename path component",
-        )?;
-
-        vec![EndpointEvent::from(
-          endpoint_events::PathComponentRenamed::from(command),
-        )]
-      }
-
-      EndpointCommand::RemovePathComponent(command) => {
-        validation.require(
-          !validation.path_component_is_root(&command.path_id),
-          "path id must not be root to remove path component",
-        )?;
-        validation.require(
-          !validation.path_component_id_exists(&command.path_id),
-          "path component must exist to remove path component",
-        )?;
-
-        vec![EndpointEvent::from(
-          endpoint_events::PathComponentRemoved::from(command),
-        )]
-      }
-
-      // Path parameters
-      // ---------------
-      EndpointCommand::AddPathParameter(command) => {
-        validation.require(
-          !validation.path_component_id_exists(&command.parent_path_id),
-          "parent path component must exist to add path parameter",
-        )?;
-        validation.require(
-          !validation.path_component_id_exists(&command.path_id),
-          "path id must be assignable to add path parameter",
-        )?;
-
-        // TODO: implement the adding of shape events as well. Probably best done
-        // from the SpecCommand handler, which can run this handler and then the
-        // shape handler separately. Using a custom command, perhaps?
-
-        vec![EndpointEvent::from(
-          endpoint_events::PathParameterAdded::from(command),
-        )]
-      }
-
-      EndpointCommand::SetPathParameterShape(command) => {
-        validation.require(
-          !validation.path_component_is_root(&command.path_id),
-          "path id can not be root to set path parameter shape",
-        )?;
-        validation.require(
-          validation.path_component_id_exists(&command.path_id),
-          "path component must exist to set path parameter shape",
-        )?;
-
-        // TODO: figure out how we can verify the shape id exists. Probably best done
-        // from the SpecCommand hanlder, which can access both projections. Perhaps by
-        // running the ShapeCommand handler first, in which we verify?
-
-        vec![EndpointEvent::from(
-          endpoint_events::PathParameterShapeSet::from(command),
-        )]
-      }
-
-      EndpointCommand::RenamePathParameter(command) => {
-        validation.require(
-          !validation.path_component_is_root(&command.path_id),
-          "path id can not be root to rename path parameter",
-        )?;
-        validation.require(
-          !validation.path_component_id_exists(&command.path_id),
-          "path component must exist to rename path parameter",
-        )?;
-
-        vec![EndpointEvent::from(
-          endpoint_events::PathParameterRenamed::from(command),
-        )]
-      }
-
-      EndpointCommand::RemovePathParameter(command) => {
-        validation.require(
-          !validation.path_component_is_root(&command.path_id),
-          "path id must not be root to remove path parameter",
-        )?;
-        validation.require(
-          !validation.path_component_id_exists(&command.path_id),
-          "path component must exist to remove path parameter",
-        )?;
-
-        vec![EndpointEvent::from(
-          endpoint_events::PathParameterRemoved::from(command),
-        )]
-      }
-
-      _ => Err(SpecCommandError::Unimplemented(
-        SpecCommand::EndpointCommand(self),
-      ))?,
-    };
-
-    Ok(events)
-  }
-}
-
-struct CommandValidationQueries<'a> {
-  command_description: String,
-  endpoint_projection: &'a EndpointProjection,
-}
-
-impl<'a> CommandValidationQueries<'a> {
-  fn require(&self, condition: bool, msg: &'static str) -> Result<(), SpecCommandError> {
-    if condition {
-      Ok(())
-    } else {
-      Err(SpecCommandError::Validation(format!(
-        "Command failed validation: {}, {:?}",
-        msg, self.command_description
-      )))
-    }
-  }
-
-  pub fn path_component_id_exists(&self, path_component_id: &PathComponentId) -> bool {
-    self
-      .endpoint_projection
-      .get_path_component_node_index(path_component_id)
-      .is_some()
-  }
-
-  pub fn path_component_is_root(&self, path_component_id: &PathComponentId) -> bool {
-    path_component_id == ROOT_PATH_ID
-  }
-}
-
-impl<'a> From<(&'a EndpointProjection, &EndpointCommand)> for CommandValidationQueries<'a> {
-  fn from(
-    (endpoint_projection, endpoint_command): (&'a EndpointProjection, &EndpointCommand),
-  ) -> Self {
-    Self {
-      command_description: format!("{:?}", endpoint_command),
-      endpoint_projection,
-    }
-  }
-}
-
-#[cfg(test)]
-mod test {
-  use super::*;
-  use serde_json::json;
-
-  #[test]
-  pub fn can_apply_events_from_endpoint_commands() {
-    let initial_events: Vec<SpecEvent> = serde_json::from_value(json!([
-      {"PathComponentAdded": {"pathId": "path_1","parentPathId": "root","name": "todos"}},
-    ]))
-    .expect("initial events should be valid spec events");
-
-    let mut initial_projection = EndpointProjection::from(initial_events);
-
-    let command: EndpointCommand = serde_json::from_value(json!(
-      { "AddPathComponent": {"pathId": "path_2", "parentPathId": "path_1", "name": "completed"}}
-    ))
-    .expect("commands should be valid endpoint events");
-
-    let new_events = initial_projection
-      .execute(command)
-      .expect("new command should be applicable to initial projection");
-
-    for event in new_events {
-      initial_projection.apply(event);
-    }
-
-    assert!(initial_projection
-      .get_path_component_node_index(&String::from("path_2"))
-      .is_some())
-  }
-}
